format comment and rename package
