+++ tools/pd-simulator/simulator/dto/stats.go
@@ -15,7 +15,7 @@ package dto
 
 import "github.com/pingcap/kvproto/pkg/pdpb"
 
-//StoreStats is to save store stats of node.
+// StoreStats is to save store stats of node.
 type StoreStats struct {
 	pdpb.StoreStats
 	ToCompactionSize uint64
